# std.crypto.pwhash.bcrypt: inline the Feistel network function