std/crypto/benchmark.zig results:

* Intel i5

```
before: 3.144 s/ops
 after: 1.922 s/ops
```

* Apple M1

```
before: 2.067 s/ops
 after: 1.373 s/ops
```